requirements.txt
numpy
pillow
cryptography
loguru
pytelegrambotapi
dotenv
flask
